TINY-8391: Initial work moving cell selection handling to model
