+++ modules/tinymce/src/models/dom/main/ts/table/Table.ts
@@ -12,11 +12,13 @@ import { TableActions } from './actions/TableActions';
 import * as Commands from './api/Commands';
 import * as Options from './api/Options';
 import * as QueryCommands from './api/QueryCommands';
+import { TableCellSelection } from './api/TableCellSelection';
 import { TableResizeHandler } from './api/TableResizeHandler';
 
 const setupTable = (editor: Editor): void => {
   Options.register(editor);
   TableResizeHandler(editor);
+  TableCellSelection(editor);
 
   const actions = TableActions(editor);
   Commands.registerCommands(editor, actions);
